Move timestamp functionality to timing module
Also move functionality to pika::chrono::detail namespace
