+++ libs/pika/timing/CMakeLists.txt
@@ -7,8 +7,17 @@
 list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/cmake")
 
 # Default location is $PIKA_ROOT/libs/timing/include
-set(timing_headers pika/timing/high_resolution_timer.hpp
-                   pika/timing/steady_clock.hpp pika/timing/tick_counter.hpp
+set(timing_headers
+    pika/timing/detail/timestamp.hpp
+    pika/timing/detail/timestamp/bgq.hpp
+    pika/timing/detail/timestamp/cuda.hpp
+    pika/timing/detail/timestamp/linux_generic.hpp
+    pika/timing/detail/timestamp/linux_x86_32.hpp
+    pika/timing/detail/timestamp/linux_x86_64.hpp
+    pika/timing/detail/timestamp/msvc.hpp
+    pika/timing/high_resolution_timer.hpp
+    pika/timing/steady_clock.hpp
+    pika/timing/tick_counter.hpp
 )
 
 # cmake will not create a separate VS project without any source files, thus
@@ -19,7 +28,6 @@ set(timing_sources)
 include(pika_add_module)
 pika_add_module(
   pika timing
-  GLOBAL_HEADER_GEN ON
   SOURCES ${timing_sources}
   HEADERS ${timing_headers}
   MODULE_DEPENDENCIES pika_config pika_hardware pika_type_support
